Some wiggling

+++ Control/InteractPerceive.h
@@ -40,8 +40,8 @@ public:
 	inline float get_move_speed() { return _move_speed; };
 	inline void set_move_speed(float speed) { _move_speed = speed; };
 	
-	inline void set_cntr_div(int div) { _cntr_div = div; }
-	inline int get_cntr_div() { return _cntr_div; }
+	inline void set_cntr_div(float div) { _cntr_div = div; }
+	inline float get_cntr_div() { return _cntr_div; }
 
 private:
 
@@ -60,6 +60,6 @@ private:
 	float _move_speed;
 	int _open_grippers_cntr;
 	ControlLogger *_logger;
-	int _cntr_div;
+	float _cntr_div;
 };
